Assetic configuration and new filters
Configuration of assetic, until now, AsseticBundle doesn't works. Now it's ok.
I've added some filters: uglifyjs and uglyfycss. This is NodeJs modules, that compact the code.

NOW, you need to install: nodejs and nodejs-legacy.

WARNING:
When you change values of css file or js file, you need to execute the command: bin/console assetic:dump, or bin/console assetic:watch, this is a service tha analyse when files are modify, and create assets on the fly.