Replace Strain delete action from Controller by a command
The deletion of a strain is very heavy, thousands of locus, features, products. We use a cascade at ORM level
then we need retrieve all of this objects in memory, and reach the PHP memory limit.

In PHP CLI, the memory limit is set very hight, then use a command is a better way.

An other way is to use a onDelete="CASCADE" and the database do the deletion, in place of PHP (ORM), but we need listen
on locus deletion event, to remove entries in ElasticSearch.